# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Eng Performance API
  version: 0.1.0
  license:
    name: Proprietary
  description: |
    Contratos de API da plataforma de performance de engenharia (Azure DevOps →
    DORA / Fluxo / IA). Este documento cobre **todos os fluxos das telas** do
    protótipo, cruzados com o catálogo de métricas e o RBAC do PRD
    (`initial-spec/initial-spec.md`).

    ## Convenções transversais
    - **Contexto de estrutura** vem no parâmetro `nodeId` (`all`, `v:<slug>`,
      `t:<slug>`, `p:<slug>`). O nível é derivável do prefixo.
    - **Frequência** (`daily|weekly|monthly`) recalcula valores e evolução; nunca
      recarrega a estrutura.
    - **Polaridade:** todo `Delta` traz `direction` (segue o sinal numérico) e
      `sentiment` (`good|bad|neutral`, segue a direção desejada da métrica). A UI
      pinta a cor por `sentiment` e a seta por `direction`.
    - **Benchmark:** `tier` (`elite|high|medium|low`) só é preenchido em métricas
      DORA; nulo nas demais.
    - **RBAC:** o backend filtra por papel. Acesso a um `nodeId` fora do escopo do
      usuário responde `403`. Comparação/ranking **de pessoas** só é retornada ao
      gestor daquele time (nunca leaderboard público nem cross-time).
servers:
  - url: /api
    description: Servido pelo próprio serviço (mesmo host da UI)
security:
  - bearerAuth: []
...
paths:
...
  /admin/integrations/azure-devops:
    get:
      tags: [Admin]
      summary: Configuração da integração com o Azure DevOps.
      responses:
        "200":
          description: Config atual (segredos redigidos).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdoIntegration" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [Admin]
      summary: Atualiza a integração (org, PAT, regra de stage de produção).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdoIntegrationInput" }
      responses:
        "200":
          description: Atualizada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdoIntegration" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /admin/ai-convention:
    get:
      tags: [Admin]
      summary: Convenção de detecção de IA (trailer/tag de commit).
      responses:
        "200":
          description: Convenção atual.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiConvention" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [Admin]
      summary: Atualiza a convenção de IA.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AiConvention" }
      responses:
        "200":
          description: Atualizada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AiConvention" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/email:
    get:
      tags: [Admin]
      summary: Configuração do servidor de e-mail (SMTP) usado no reset de senha.
      responses:
        "200":
          description: Config atual. `passwordSet` indica se há senha salva; a senha em si nunca é retornada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailSettings" }
        "403": { $ref: "#/components/responses/Forbidden" }
    put:
      tags: [Admin]
      summary: Atualiza a configuração de e-mail.
      description: |
        Omitir (ou enviar em branco) `password` preserva a senha já guardada. Sem
        um SMTP habilitado com host, o envio cai no fallback de log.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmailSettingsInput" }
      responses:
        "200":
          description: Atualizada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmailSettings" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /admin/email/test:
    post:
      tags: [Admin]
      summary: Envia um e-mail de teste com a configuração atual.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TestEmailRequest" }
      responses:
        "204": { description: Enviado (ou registrado no log, se sem SMTP). }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502":
          description: Falha na entrega pelo servidor SMTP configurado.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: Autenticado, mas sem permissão para este nó/recurso (RBAC).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ValidationError:
      description: Corpo inválido.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    AdoIntegration:
      type: object
      required: [organizationUrl, connected]
      properties:
        organizationUrl: { type: string, format: uri }
        connected: { type: boolean }
        patRedacted: { type: string, description: PAT mascarado (nunca retornado em claro). }
        productionStageRule:
          type: string
          description: Regra global de stage de produção (times podem sobrescrever).
        lastValidatedAt: { type: string, format: date-time, nullable: true }
...
    AdoIntegrationInput:
      type: object
      required: [organizationUrl, personalAccessToken]
      properties:
        organizationUrl: { type: string, format: uri }
        personalAccessToken: { type: string, format: password }
        productionStageRule: { type: string }
    AiConvention:
      type: object
      required: [strategy]
      properties:
        strategy: { type: string, enum: [trailer, tag] }
        trailer:
          type: string
          example: "Co-authored-by: Copilot"
        tag: { type: string, example: "[ai]" }
        caseSensitive: { type: boolean, default: false }
    EmailSettings:
      type: object
      required: [enabled, transport, passwordSet]
      properties:
        enabled: { type: boolean }
        host: { type: string, nullable: true }
        port: { type: integer, nullable: true }
        transport: { type: string, enum: [none, starttls, ssl] }
        username: { type: string, nullable: true }
        passwordSet:
          type: boolean
          description: Indica se há senha salva — a senha em si nunca é retornada.
        fromAddress: { type: string, format: email, nullable: true }
        fromName: { type: string, nullable: true }
        appBaseUrl:
          type: string
          format: uri
          nullable: true
          description: Base usada para montar o link de reset de senha no e-mail.
    EmailSettingsInput:
      type: object
      required: [enabled, transport]
      properties:
        enabled: { type: boolean }
        host: { type: string }
        port: { type: integer }
        transport: { type: string, enum: [none, starttls, ssl] }
        username: { type: string }
        password:
          type: string
          format: password
          description: Omitir (ou enviar em branco) preserva a senha já guardada.
        fromAddress: { type: string, format: email }
        fromName: { type: string }
        appBaseUrl: { type: string, format: uri }
    TestEmailRequest:
      type: object
      required: [to]
      properties:
        to: { type: string, format: email }
...
    Problem:
      type: object
      description: RFC 7807 problem detail.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }